+++ packages/orq-rc/src/orq_ai_sdk/models/bulkcreatedatapointsop.py
@@ -387,5 +387,5 @@ class ResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2025-02-27T08:02:43.136Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2025-02-27T10:11:59.979Z")
     r"""The date and time the resource was last updated"""